+++ api/src/container/IOC.config.ts
@@ -149,11 +149,12 @@ import GenericError from '../errors/GenericError';
 import { LogPriority } from '../enums/LogPriority';
 import { Mode } from '../enums/Mode';
 import AwsKeyStoreConfig from '../backends/configuration/AwsKeyStoreConfig';
-import Context = interfaces.Context;
-import Factory = interfaces.Factory;
 import AccountService from '../accounts/AccountService';
 import KinesisService from '../aws/KinesisService';
 import SignUpService from '../signup/SignUpService';
+import OrgService from '../orgs/OrgService';
+import Context = interfaces.Context;
+import Factory = interfaces.Factory;
 
 const container = new Container();
 
@@ -234,6 +235,7 @@ container.bind<Route53Service>(TYPES.Route53Service).to(Route53Service).inSingle
 container.bind<StripeService>(TYPES.StripeService).to(StripeService).inSingletonScope();
 container.bind<SignUpService>(TYPES.SignUpService).to(SignUpService).inSingletonScope();
 container.bind<AccountService>(TYPES.AccountService).to(AccountService).inSingletonScope();
+container.bind<OrgService>(TYPES.OrgService).to(OrgService).inSingletonScope();
 container.bind<ConsoleLogger>(TYPES.ConsoleLogger).to(ConsoleLogger).inSingletonScope();
 
 [
