+++ api/src/jobs/StripeSync.ts
@@ -2,9 +2,9 @@ import container from '../container/IOC.config';
 import TYPES from '../container/IOC.types';
 import RepoFromModelFactory from '../container/factoryTypes/RepoFromModelFactory';
 import Org from '../mongo/models/Org';
-import OrgManager from '../managers/OrgManager';
+import OrgService from '../orgs/OrgService';
 
-const orgManager = container.get<OrgManager>(TYPES.OrgManager);
+const orgService = container.get<OrgService>(TYPES.OrgService);
 const repoFactory = container.get<RepoFromModelFactory>(TYPES.RepoFromModelFactory);
 
 const DELAY_MS = 500;
@@ -13,7 +13,7 @@ export const stripeSync = {
     name: 'StripeSync',
     job: async () => {
         for await (const org of repoFactory(Org).findIterator({})) {
-            await orgManager.updateOrgBillingCycle(org);
+            await orgService.updateOrgBillingCycle(org);
             //sleep between calls...
             await new Promise<void>((resolve) => setTimeout(() => resolve(), DELAY_MS));
         }
